Separate implementation from header
This allows incremental build.
Also, implicit rules are used in the Makefile.
