+++ symbol.cpp
@@ -1,0 +1,22 @@
+#include "symbol.hpp"
+
+#include <map>
+#include <memory>
+#include <string>
+#include <utility>
+
+std::shared_ptr<SymbolEntry> SymbolTable::Add(
+    std::unique_ptr<SymbolEntry> entry) {
+  const std::string& id = entry->id;  // to reference id after moved
+  if (!Probe(id)) {
+    entries_.insert({id, std::shared_ptr<SymbolEntry>{std::move(entry)}});
+  }
+  return entries_.at(id);
+}
+
+std::shared_ptr<SymbolEntry> SymbolTable::Probe(const std::string& id) const {
+  if (entries_.count(id)) {
+    return entries_.at(id);
+  }
+  return nullptr;
+}
