+++ symbol.hpp
@@ -4,7 +4,6 @@
 #include <map>
 #include <memory>
 #include <string>
-#include <utility>
 
 #include "type.hpp"
 
@@ -21,20 +20,9 @@ class SymbolTable {
   /// already in the table.
   /// @returns The added entry if the `id` of the `entry` isn't
   /// already in the table; otherwise, the original entry.
-  std::shared_ptr<SymbolEntry> Add(std::unique_ptr<SymbolEntry> entry) {
-    const std::string& id = entry->id;  // to reference id after moved
-    if (!Probe(id)) {
-      entries_.insert({id, std::shared_ptr<SymbolEntry>{std::move(entry)}});
-    }
-    return entries_.at(id);
-  }
+  std::shared_ptr<SymbolEntry> Add(std::unique_ptr<SymbolEntry> entry);
 
-  std::shared_ptr<SymbolEntry> Probe(const std::string& id) const {
-    if (entries_.count(id)) {
-      return entries_.at(id);
-    }
-    return nullptr;
-  }
+  std::shared_ptr<SymbolEntry> Probe(const std::string& id) const;
 
  private:
   std::map<std::string, std::shared_ptr<SymbolEntry>> entries_{};
